feat: extract DummyProofVerification to primitives
