+++ pallets/market/Cargo.toml
@@ -39,6 +39,7 @@ env_logger = { workspace = true }
 pallet-balances = { workspace = true, default-features = false }
 pallet-proofs = { workspace = true, default-features = false }
 pallet-storage-provider = { workspace = true, default-features = false }
+primitives = { workspace = true, default-features = false, features = ["builder", "testing"] }
 sp-core = { workspace = true, default-features = false }
 sp-io = { workspace = true }
 sp-runtime = { workspace = true, default-features = false }
@@ -49,6 +50,8 @@ runtime-benchmarks = [
   "frame-benchmarking/runtime-benchmarks",
   "frame-support/runtime-benchmarks",
   "frame-system/runtime-benchmarks",
+  "primitives/builder",
+  "primitives/testing",
   "sp-runtime/runtime-benchmarks",
 ]
 std = [
