+++ pallets/storage-provider/Cargo.toml
@@ -38,7 +38,7 @@ env_logger = { workspace = true }
 multihash-codetable = { workspace = true, features = ["blake2b"] }
 pallet-balances = { workspace = true, default-features = false }
 pallet-market = { workspace = true, default-features = false }
-primitives = { workspace = true, default-features = false, features = ["builder"] }
+primitives = { workspace = true, default-features = false, features = ["builder", "testing"] }
 rstest = { workspace = true }
 sp-io = { workspace = true }
 
@@ -48,6 +48,7 @@ runtime-benchmarks = [
   "frame-benchmarking/runtime-benchmarks",
   "frame-support/runtime-benchmarks",
   "frame-system/runtime-benchmarks",
+  "primitives/testing",
   "sp-runtime/runtime-benchmarks",
 ]
 std = [
