+++ pallets/storage-provider/src/tests/submit_windowed_post.rs
@@ -1,5 +1,5 @@
 use frame_support::{assert_noop, assert_ok};
-use primitives::PartitionNumber;
+use primitives::{testing::INVALID_PROOF, PartitionNumber};
 use rstest::rstest;
 use sp_core::bounded_vec;
 use sp_runtime::{BoundedVec, DispatchError};
@@ -12,7 +12,7 @@ use crate::{
         account, declare_faults::setup_sp_with_many_sectors_multiple_partitions, events,
         new_test_ext, register_storage_provider, run_to_block, DealProposalBuilder, Market,
         RuntimeEvent, RuntimeOrigin, SectorPreCommitInfoBuilder, StorageProvider,
-        SubmitWindowedPoStBuilder, System, Test, ALICE, BOB, INVALID_PROOF,
+        SubmitWindowedPoStBuilder, System, Test, ALICE, BOB,
     },
     Config,
 };
